YD-235 Fixed locale handling for admin service
* Simplified the localization filter by removing the locale-specific resource loading, as the app developers anyway cannot use that
* Moved the localization filter to core, so it's also applied to the admin service
* Added a small localization test to the admin service
